requirements.txt
aiofiles==0.7.0
aiosqlite==0.16.1
anyio==3.2.1
asgiref==3.3.4
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.1.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
ecdsa==0.17.0
environs==9.5.0
fastapi==0.65.2
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==3.2
inflect==5.3.0
iniconfig==1.1.1
iso8601==0.1.14
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.17.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pyparsing==3.0.9
pypika-tortoise==0.1.1
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
requests==2.28.1
rfc3986==1.5.0
rich==10.4.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
Telethon==1.24.0
tomli==2.0.1
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
urllib3==1.26.10
uvicorn==0.14.0
wincertstore==0.2
